// src/types/djradio.d.ts
export type djRadiosType = {
  id: number,
  picUrl: string,
  name: string,
  programCount: number,
  subCount: number,
  userId: number,
  nickname: string
};

export type radiosType = {
  "dj": {
    "defaultAvatar": boolean,
    "province": number,
    "authStatus": number,
    "followed": boolean,
    "avatarUrl": string,
    "accountStatus": number,
    "gender": number,
    "city": number,
    "birthday": number,
    "userId": number,
    "userType": number,
    "nickname": string,
    "signature": string,
    "description": string,
    "detailDescription": string,
    "avatarImgId": number,
    "backgroundImgId": number,
    "backgroundUrl": string,
    "authority": number,
    "mutual": boolean,
    "expertTags": null,
    "experts": null,
    "djStatus": number,
    "vipType": number,
    "remarkName": string | null,
    "avatarImgIdStr": string,
    "backgroundImgIdStr": string,
    "avatarImgId_str": string
  },
  "category": string,
  "buyed": boolean,
  "price": number,
  "originalPrice": number,
  "discountPrice": string | null,
  "purchaseCount": number,
  "lastProgramName": string,
  "videos": string | null,
  "finished": boolean,
  "underShelf": boolean,
  "liveInfo": string | null,
  "categoryId": number,
  "picId": number,
  "createTime": number,
  "lastProgramId": number,
  "feeScope": number,
  "subCount": number,
  "programCount": number,
  "lastProgramCreateTime": number,
  "radioFeeType": number,
  "picUrl": string,
  "desc": string,
  "name": string,
  "id": number,
  "rcmdtext": string
};